requirements.txt
asgiref==3.9.1
Django==5.2.3
gunicorn==23.0.0
packaging==25.0
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
django-import-export
